First implementation of cuda completed. The following is implemented in CUDA 1) wfsgrad ray tracing 2) performance evaluation, RMS OPD part To be implemented 1) wfsgrad gradient calculation 2) performance evaluation PSF computation 3) tomography/fit

+++ lib/imat.c
@@ -26,7 +26,7 @@ imat* inew(long nx, long ny){
     imat *A=calloc(1, sizeof(imat));
     A->nx=nx;
     A->ny=ny;
-    A->p=calloc(nx*ny, sizeof(int));
+    A->p=calloc(nx*ny, sizeof(long));
     return A;
 }
